fix: split assembler.rs into types + impl files to satisfy ≤500 LOC invariant
